功能完善

+++ include/common/smd_defines.h
@@ -1,7 +1,7 @@
 ﻿#pragma once
 
 namespace smd {
-using SMD_POINTER = unsigned long;
+using SMD_POINTER = unsigned long long;
 const SMD_POINTER SMD_NULL_PTR = 0;
 
 }
